+++ changelog.md
@@ -1,6 +1,8 @@
 ﻿# GeoTagNinja Changelog
 v0.6.83XX [2022XXXX]
 - Finally managed to get hold of ReSharper so the code has been refactored in totality. Famous last words but it shouldn't affect usage.
+- Bugfixes:
+-- If a user's "Pictures" folder had been moved and renamed the could would break because Windows treats special folders in odd ways. (e.g. if the Pics folder is called "Pics" Windows would still show "Pictures", which doesn't per se exist.)
 
 v0.6.8350 [20221111]
 - NEW FEATURE: (experimental for now) -> sync/import GPS Track Files.